Pixels

Description

Modulo : Pixels Descripcion : Representacion de pixeles prendidos y apagados en terminal Autores : Maria Bracamonte, 10-11147 Carlos Perez, 13-11089 Souce : https:github.comCapf96Proyecto1-Lenguajes Modulo que permite la impresion en pantalla de caracteres ASCII en una representacion sencilla de pixeles en blanco y negro. Los pixeles se componen de una lista con 7 elementos normalmente, donde cada elemento representa una linea de la impresion en pantalla. Cada pixel dentro de un tipo Pixels esta representado por un caracter, que puede ser * que representa el pixel prendido, y ' ' que representa el pixel apagado. A su vez definimos una serie de funciones que trabajan con este nuevo tipo de dato, realizando operaciones normalmente encontrado en una pantalla de pixeles. Tambien definimos funciones para transformar dicho Pixels en un String.

Documentation

type Pixels = [String]

El tipo de dato Pixels representa una impresion en pantalla de algun caracter o String a traves de pixeles prendidos y apagados representados por * y ' ' respectivamente. Esta compuesto de 7 elementos normalmente, los cuales representan una linea del Pixels.

font :: Char -> Pixels

La funcion font toma un caracter y este lo transforma en un tipo pixel a tipo Pixels, buscando el caracter introducido en la lista fontBitmap, en donde, en caso de ser encontrado, trabaja sobre la lista proporcionada que contiene el valor hexadecimal de los bits prendidos y apagados de cada columna del pixel; en caso de no ser encontrado devuelve un Pixels con todos los pixeles prendidos.

showPixels :: Pixels -> IO ()

La funcion showPixels recibe un Pixels y lo imprime en pantalla de tal manera que forme el caracter que esta representando.

pixelsToString :: Pixels -> String

La funcion pixelsToString recibe un Pixels y lo transforma en un String con debidos saltos de linea para su correcta representacion.

pixelListToPixels :: [Pixels] -> Pixels

La funcion pixelListToPixels recibe una lista de Pixels y une todos los elementos de esta con un string vacio de separacion entre estos.

pixelListToString :: [Pixels] -> String

La funcion pixelListToString recibe una lista de Pixels y los transforma en un String.

concatPixels :: [Pixels] -> Pixels

La funcion concatPixels recibe una lista de pixeles y los concatena entre ellos formando un solo Pixels.

messageToPixels :: String -> Pixels

La funcion messageToPixels recibe un String y lo transforma en un elemento de tipo Pixels.

up :: Pixels -> Pixels

La funcion up recibe un Pixels y toma el primer elemento de este y lo coloca en la ultima posicion

down :: Pixels -> Pixels

La funcion down recibe un Pixels y toma el ultimo elemento de este y lo coloca de primero

left :: Pixels -> Pixels

La funcion left recibe un Pixels y a toma el primer elemento de cada elemento que lo compone y lo coloca de ultimo

right :: Pixels -> Pixels

La funcion left recibe un Pixels y a toma el ultimo elemento de cada elemento que lo compone y lo coloca de primero

upsideDown :: Pixels -> Pixels

La funcion upsideDown recibe un Pixels e intercambia todos los elementos que lo componen tal que el primero es el ultimo, el segundo el penultimo, etc. De esta manera al realizar la impresion del resultado imprime el Pixels de cabeza

backwards :: Pixels -> Pixels

La funcion backwards recibe un Pixels y voltea la posicion de cada uno de los elementos de sus elementos tal que el primero es el ultimo, el segundo el penultimo, etc. De esta forma, el resultado de la impresion del Pixels seria la version espejo del Pixels original.

negative :: Pixels -> Pixels

La funcion negative recibe un Pixels y transforma todos los pixeles prendidos a apagados y todos los pixeles apagados a prendidos, es decir, todos los * en ' ' y viceversa.